pyproject.toml
[tool.poetry]
name = "mon-precos-dbt-sql"
version = "0.1.0"
description = ""
authors = ["Giovane Carvalho <degio2009@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
yfinance = "^0.2.48"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
